% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moran}
\alias{moran}
\title{Moran's Index in TNBC, based on Gabriel neighborhoods}
\format{
csv with 41 patients, 41 Moran's I scores
\describe{
\item{p_id}{patient ID}
\item{cellSize}{Moran's Index for cell size}
\item{Vimentin}{Moran's Index for the Vimentin expression}
}
}
\usage{
moran
}
\description{
Computed from the Mibi-TOF image data in
"A Structured Tumor-Immune Microenvironment in
Triple Negative Breast Cancer Revealed by
Multiplexed Ion Beam Imaging" (Keren et al., Cell 2018)
}
\details{
Output of the compute_morans.Rmd script
}
\keyword{datasets}
